pyproject.toml
[build-system]
requires = ["setuptools>=42", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "docx2md"
version = "0.1.0"
description = "Convert Microsoft Word documents with complex math to Markdown+LaTeX"
readme = "README.md"
authors = [
    {name = "ThreeFold", email = "info@threefold.io"}
]
license = {text = "Apache-2.0"}
classifiers = [
    "Development Status :: 3 - Alpha",
    "Intended Audience :: Science/Research",
    "Intended Audience :: Education",
    "License :: OSI Approved :: Apache Software License",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.8",
    "Programming Language :: Python :: 3.9",
    "Programming Language :: Python :: 3.10",
    "Programming Language :: Python :: 3.11",
    "Topic :: Text Processing :: Markup",
    "Topic :: Text Processing :: Markup :: Markdown",
    "Topic :: Text Processing :: Markup :: LaTeX",
    "Topic :: Office/Business :: Office Suites",
]
keywords = ["markdown", "latex", "word", "docx", "conversion", "math", "equations"]
requires-python = ">=3.8"
dependencies = [
    "python-docx==1.0.0",
    "pypandoc>=1.11",
    "click>=8.1.3",
    "pyyaml>=6.0",
    "rich>=13.3.5",
    "pillow>=9.5.0",
    "regex>=2023.3.23",
]

[project.optional-dependencies]
dev = [
    "pytest>=7.3.1",
    "pytest-cov>=4.1.0",
    "black>=23.3.0",
    "isort>=5.12.0",
    "flake8>=6.0.0",
    "mypy>=1.3.0",
    "build>=0.10.0",
]

[project.urls]
"Homepage" = "https://github.com/ucli-tools/docx2md"
"Bug Tracker" = "https://github.com/ucli-tools/docx2md/issues"
"Documentation" = "https://github.com/ucli-tools/docx2md#readme"

[project.scripts]
docx2md = "docx2md.cli:main"

[tool.setuptools]
packages = ["docx2md"]

[tool.black]
line-length = 88
target-version = ["py38", "py39", "py310", "py311"]
include = '\.pyi?$'

[tool.isort]
profile = "black"
line_length = 88

[tool.mypy]
python_version = "3.8"
warn_return_any = true
warn_unused_configs = true
disallow_untyped_defs = true
disallow_incomplete_defs = true

[tool.pytest.ini_options]
testpaths = ["tests"]
python_files = "test_*.py"
python_functions = "test_*"
